+++ src/scanner/scanner.module.ts
@@ -1,9 +1,11 @@
 import { Module } from '@nestjs/common';
 
+import { ScannerController } from './scanner.controller';
 import { ScannerGateway } from './scanner.gateway';
 import { ScannerService } from './scanner.service';
 
 @Module({
+  controllers: [ScannerController],
   providers: [ScannerService, ScannerGateway],
   exports: [ScannerService],
 })